More flexibility in controlling the download behaviour:
Allow passing an NSURLRequest instead of an NSURL.
Allow passing a destination file to download directly to it, instead of using a temporary file.
Allow pausing a connection - so you can resume it later on the same file.
Allow a standalone DGDownloadManagerFile which is not involved in the queue management, concurrent limit etc.